chore(checker): move Checker struct to src/checker.rs
